requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.2
opencv-python==4.5.1.48
pyjwt==2.0.0
python-dotenv==0.15.0
pyzbar==0.1.8
requests==2.25.1
six==1.15.0
urllib3==1.26.2
wincertstore==0.2
py2neo==2021.1.1